Revert "-Aggiunto sistema di registrazione"
This reverts commit 4b1b3d5e166cd158b84219b790bc9c4af0be434d.

+++ Sketchbook/src/model/ProductModelDS.java
@@ -81,7 +81,7 @@ public class ProductModelDS implements ProductModel {
 				bean.setID(rs.getInt("ID"));
 				bean.setName(rs.getString("NOME"));
 				bean.setDescription(rs.getString("DESCRIZIONE"));
-				bean.setPrice(rs.getDouble("PREZZO"));
+				bean.setPrice(rs.getInt("PREZZO"));
 				bean.setQuantity(rs.getInt("QUANTITA"));
 			}
 
@@ -150,7 +150,7 @@ public class ProductModelDS implements ProductModel {
 				bean.setID(rs.getInt("ID"));
 				bean.setName(rs.getString("NOME"));
 				bean.setDescription(rs.getString("DESCRIZIONE"));
-				bean.setPrice(rs.getDouble("PREZZO"));
+				bean.setPrice(rs.getInt("PREZZO"));
 				bean.setQuantity(rs.getInt("QUANTITA"));
 				products.add(bean);
 			}
